refactor(client): move driver introspection into jumpstarter.client for CLI/MCP parity
Move walk_click_tree, list_drivers, get_driver_methods and their helpers
from jumpstarter_mcp.introspect into jumpstarter.client.introspect so the
CLI (e.g. a future 'jmp describe lease --devices') can share them with the
MCP server. Add describe_client plus describe_devices/describe_devices_async
helpers that attach to an existing lease (never creating or releasing it)
and return a plain-serializable devices dict. jumpstarter_mcp.introspect
remains as a backward-compatible re-export shim.
